🔨 refactor: Separating qna and faq
